Fix: Extract AbstractDifferTestCase
